test(validation): pin the heading boundary and wrapper failure branch
The diff-parsing class moves to its own file at the file-size ceiling,
along the class seam; the heading-to-body case and the boolean
wrapper's git-failure return each gain a discriminating test.

Refs #5337
